v0.2.24: Cross-architecture hardening complete
Phase 1A — UART driver abstraction:
  serial.hpp pure interface, x86_64 impl moved to serial.cpp

Phase 1B — Cross-arch atomics:
  kernel::atomic<T> wrapper, 12 __sync_synchronize() replaced,
  __atomic_* wrapped in spinlock/spsc/ring_buffer/dmesg

Phase 2C — Boot flow unification:
  libfdt subset ported (6 files), BootInfo struct,
  higherhalf_entry unified, FDT memory parsing for aarch64/riscv64

Phase 2D — Virtio transport unification:
  unified PCI HAL (CF8/CFC + ECAM), shared virtio_pci.cpp,
  virtio-blk/net unguarded, probed on all arches

Phase 3E — Memory model tests (12 atomic tests)
Phase 3F — Cross-arch test suite (16 shared tests)
Phase 3G — Renode CI gate (make renode-test for aarch64/riscv64)

693/693 tests PASS (was 665), 0 FAILED

+++ src/kernel/arch/pci.cpp
@@ -17,7 +17,7 @@
  */
 
 /// @file pci.cpp
-/// @brief PCI bus enumeration using CF8/CFC config space access.
+/// @brief PCI bus enumeration — arch-independent (uses HAL config space access).
 
 #include <kernel/arch/pci.hpp>
 #include <logger.hpp>
@@ -388,7 +388,7 @@ PciDeviceInfo pci_read_device_info(PciBdf bdf) {
     info.bdf = bdf;
     info.vendor_id   = pci_read_vendor(bdf);
     info.device_id   = pci_read_device(bdf);
-    uint32_t addr_base = pci_make_addr(bdf, 0);
+    uint64_t addr_base = pci_make_addr(bdf, 0);
     info.revision    = pci_config_readb(addr_base + PCI_REVISION);
     info.prog_if     = pci_config_readb(addr_base + PCI_PROG_IF);
     info.subclass    = pci_config_readb(addr_base + PCI_SUBCLASS);
